changed datatype of starttime storages from int to long (2-4 bytes)

+++ fanControl.ino
@@ -12,9 +12,9 @@ enum FANSTATE {
   STOPPED
 };
 FANSTATE fanState = STOPPED;
-int fanSpinStartTime;
-int fanRollStartTime;
-int fanStopStartTime;
+long fanSpinStartTime;
+long fanRollStartTime;
+long fanStopStartTime;
 String message;
 
 void setup() {
